requirements.txt
absl-py==0.3.0
apache-libcloud==2.4.0
astor==0.7.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
click==6.7
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.2
Flask==1.0.2
future==0.17.1
futures==3.1.1
gast==0.2.0
grpcio==1.14.1
idna==2.8
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.15.0
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.0
py-tes==0.3.0
pyparsing==2.2.0
pytesseract==0.2.6
python-dateutil==2.6.1
pytz==2018.5
requests==2.21.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
